migrate test1 and test2 to gtest
