+++ deque/.gitignore
@@ -1,0 +1,17 @@
+CMakeCache.txt
+CMakeFiles
+CMakeScripts
+Testing
+Makefile
+cmake_install.cmake
+install_manifest.txt
+compile_commands.json
+CTestTestfile.cmake
+
+cmake-build*/
+Debug/
+Release/
+Bin/
+
+.idea/
+
