--- deque/data/class-integer.hpp
@@ -1,11 +1,0 @@
-class Integer {
-private:
-	int data;
-public:
-	Integer(const int &value) : data(value) {}
-	Integer(const Integer &other) : data(other.data) {}
-	bool operator==(const Integer &t)
-	{
-		return data == t.data;
-	}
-};
