+++ deque/test_main.cpp
@@ -1,0 +1,10 @@
+//
+// Created by Alex Chi on 2019-07-22.
+//
+
+#include "gtest/gtest.h"
+
+int main(int argc, char **argv) {
+    ::testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
+}
